+++ src/nsidc/metgen/readers/csv.py
@@ -13,7 +13,7 @@ from nsidc.metgen.readers import utilities
 
 def extract_metadata(
     csv_path: str,
-    premet_content: dict,
+    temporal_content: list,
     spatial_content: list,
     configuration: Config,
     _,
@@ -21,14 +21,14 @@ def extract_metadata(
     df = pd.read_csv(csv_path)
 
     return {
-        "temporal": data_datetime(df, premet_content),
+        "temporal": data_datetime(df, temporal_content),
         "geometry": bbox(spatial_values(df, spatial_content, configuration)),
     }
 
 
-def data_datetime(df, premet_content: dict) -> list:
-    if premet_content:
-        return utilities.temporal_from_premet(premet_content)
+def data_datetime(df, temporal_content: list) -> list:
+    if temporal_content:
+        return temporal_content
 
     def formatted(date, dt):
         return (
@@ -41,10 +41,12 @@ def data_datetime(df, premet_content: dict) -> list:
     data_times = df["TIME"]
 
     if data_dates.size > 0 and data_times.size > 0:
-        return [
-            formatted(data_dates.iat[0], data_times.iat[0]),
-            formatted(data_dates.iat[-1], data_times.iat[-1]),
-        ]
+        return utilities.refine_temporal(
+            [
+                formatted(data_dates.iat[0], data_times.iat[0]),
+                formatted(data_dates.iat[-1], data_times.iat[-1]),
+            ]
+        )
     else:
         return None